# cli: Add OOTB TypeScript support and big build performance improvements