Fixes #7

Well, the goal is to provide OOTB TypeScript support without having to add extra configuration to get started. The implementation is also not opinionated about the settings so we use `@babel/preset-typescript` which does not do type checking at build time and will not degrade build time performance.

As part of the support also adding some cache level settings for webpack and for babel itself, this brings considerably decreases the build time for larger projects and medium projects, decreasing around 11x in some cases 🎉.

To show some examples, considering some examples from our [collection of examples](https://github.com/fleetfn/examples):

| Example        | Old           | New  | Gain |
| ------------- |-------------|-------------|-------------|
| faunadb      | 9s | 769ms | 11x |
| simple-http-endpoint | 1s      |    655ms | 1.4x |

It is considered that projects that have more dependencies or more Functions are the big beneficiaries because very simple functions like an echo function despite having a very good improvement does not have much effect.